+++ common/src/main/java/club/iananderson/seasonhud/platform/services/IPlatformHelper.java
@@ -3,7 +3,7 @@ package club.iananderson.seasonhud.platform.services;
 public interface IPlatformHelper {
 
     /**
-     * Gets the name of the current platform
+     * Gets the name of the current platform.
      *
      * @return The name of the current platform.
      */