+++ fabric/src/main/java/club/iananderson/seasonhud/client/minimaps/JourneyMap.java
@@ -16,11 +16,11 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 
 import static club.iananderson.seasonhud.Common.SEASON_STYLE;
-import static club.iananderson.seasonhud.SeasonHUD.MOD_ID;
+import static club.iananderson.seasonhud.Common.MOD_ID;
 import static club.iananderson.seasonhud.config.Config.*;
-import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.loadedMinimap;
-import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.shouldDraw;
-import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonName;
+import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.minimapLoaded;
+import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.shouldDrawMinimapHud;
+import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonHudName;
 
 public class JourneyMap implements HudRenderCallback{
     public static JourneyMap HUD_INSTANCE;
@@ -32,8 +32,8 @@ public class JourneyMap implements HudRenderCallback{
 
     private static String getSeason(){
         MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
-                getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
-                getSeasonName().get(1).copy());
+                getSeasonHudName().get(0).copy().withStyle(SEASON_STYLE),
+                getSeasonHudName().get(1).copy());
 
         return seasonCombined.getString();
     }
@@ -42,15 +42,15 @@ public class JourneyMap implements HudRenderCallback{
     public void onHudRender(GuiGraphics seasonStack, float alpha) {
         Minecraft mc = Minecraft.getInstance();
         MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
-                getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
-                getSeasonName().get(1).copy());
+                getSeasonHudName().get(0).copy().withStyle(SEASON_STYLE),
+                getSeasonHudName().get(1).copy());
 
         if(Services.PLATFORM.isModLoaded("journeymap") && !enableMod.get()) {
             ThemeLabelSource.InfoSlot Season = ThemeLabelSource.create(MOD_ID,"menu.seasonhud.infodisplay.season",1000L,1L, JourneyMap::getSeason);
             // Should only show up if the "Enable Mod" option in the SeasonHUD menu/config is disabled. Icon currently doesn't work
         }
 
-        if (loadedMinimap("journeymap")) {
+        if (minimapLoaded("journeymap")) {
             DisplayVars vars = UIManager.INSTANCE.getMiniMap().getDisplayVars();
             JourneymapClient jm = JourneymapClient.getInstance();
 
@@ -88,7 +88,7 @@ public class JourneyMap implements HudRenderCallback{
             int startX = (int) (textureX + halfWidth);
             int startY = (int) (textureY + (journeyMapAboveMap.get() ? - margin - labelHeight : minimapHeight + margin));
 
-            if (shouldDraw() && jm.getActiveMiniMapProperties().enabled.get()) {
+            if (shouldDrawMinimapHud() && jm.getActiveMiniMapProperties().enabled.get()) {
                 int labelX = (int) startX;
                 int labelY = startY + (journeyMapAboveMap.get() ? -topLabelHeight : bottomLabelHeight);