+++ fabric/src/main/java/club/iananderson/seasonhud/impl/minimaps/JourneyMapAPI.java
@@ -1,6 +1,6 @@
 package club.iananderson.seasonhud.impl.minimaps;
 
-import club.iananderson.seasonhud.SeasonHUD;
+import club.iananderson.seasonhud.Common;
 import journeymap.client.JourneymapClient;
 import journeymap.client.api.IClientAPI;
 import journeymap.client.api.IClientPlugin;
@@ -27,7 +27,7 @@ public class JourneyMapAPI implements IClientPlugin {
 
     @Override
     public String getModId() {
-        return SeasonHUD.MOD_ID;
+        return Common.MOD_ID;
     }
 
     @Override