+++ fabric/src/main/java/club/iananderson/seasonhud/impl/minimaps/XaeroInfoDisplays.java
@@ -1,6 +1,5 @@
 package club.iananderson.seasonhud.impl.minimaps;
 
-import club.iananderson.seasonhud.config.Config;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import xaero.common.minimap.info.InfoDisplay;
@@ -11,9 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static club.iananderson.seasonhud.Common.SEASON_STYLE;
-import static club.iananderson.seasonhud.impl.seasons.Calendar.calendar;
-import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonName;
-import static club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.dimensionHideHUD;
+import static club.iananderson.seasonhud.impl.seasons.CurrentSeason.getSeasonHudName;
 
 public class XaeroInfoDisplays {
     private static final List<InfoDisplay<?>> ALL = new ArrayList<>();
@@ -21,11 +18,15 @@ public class XaeroInfoDisplays {
 
     static{
         SEASON = new InfoDisplay<>("season", Component.translatable("menu.seasonhud.infodisplay.season"), true, InfoDisplayCommonStateCodecs.BOOLEAN, InfoDisplayCommonWidgetFactories.OFF_ON, (displayInfo, compiler, session, processor, x, y, w, h, scale, size, playerBlockX, playerBlockY, playerBlockZ, playerPos) -> {
-            MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
-                    getSeasonName().get(0).copy().withStyle(SEASON_STYLE),
-                    getSeasonName().get(1).copy());
+            if(getSeasonHudName().isEmpty()){
+                return;
+            }
 
-            if ((Boolean)displayInfo.getState() && !dimensionHideHUD() && calendar() && Config.enableMod.get() && Config.enableMinimapIntegration.get()) {
+            MutableComponent seasonCombined = Component.translatable("desc.seasonhud.combined",
+                    getSeasonHudName().get(0).copy().withStyle(SEASON_STYLE),
+                    getSeasonHudName().get(1).copy());
+
+            if ((Boolean) displayInfo.getState() && CurrentMinimap.shouldDrawMinimapHud()) {
                 compiler.addLine(seasonCombined);
             }
         },ALL);